backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
beautifulsoup4
readability-lxml
lxml
markdownify
sentence-transformers
faiss-cpu
rank-bm25
numpy
pandas
openpyxl
python-calamine
jsonschema
python-multipart
tqdm
tenacity
python-dotenv
